+++ src/components/Search/Search.tsx
@@ -4,11 +4,10 @@ import style from './Search.module.scss';
 interface SearchProps {
   value: string;
   onSearchChange: (value: string) => void;
-  // TODO: переименовать
-  handleClick: (e: React.KeyboardEvent<HTMLInputElement>) => void;
+  onKeyDown: (e: React.KeyboardEvent<HTMLInputElement>) => void;
 }
 
-export const Search: React.FC<SearchProps> = ({ value, onSearchChange, handleClick }) => {
+export const Search: React.FC<SearchProps> = ({ value, onSearchChange, onKeyDown }) => {
   return (
     <div className={style.form_container}>
       <form className={style.search_form}>
@@ -17,7 +16,7 @@ export const Search: React.FC<SearchProps> = ({ value, onSearchChange, handleCli
           placeholder={'Search in the anime...'}
           className={style.search__input}
           onChange={(event) => onSearchChange(event.target.value)}
-          onKeyDown={(event) => handleClick(event)}
+          onKeyDown={(event) => onKeyDown(event)}
           value={value}
         />
         <img src={img} alt="img" className={style.search__img} />
