+++ src/const/api-query.ts
@@ -1,1 +1,2 @@
 export const ANIME_API_ALL = 'https://shikimori.one/api';
+export const ITEM_ON_PAGE = 16;
