+++ src/hooks/index.ts
@@ -1,5 +1,4 @@
 export * from './use-mount';
 export * from './use-unmount';
-export * from './use-home';
 export * from './use-modal';
 export * from './use-http';
