+++ src/utils/index.ts
@@ -1,4 +1,3 @@
-export * from './api-default-value';
 export * from './get-current-page';
 export * from './get-local';
 export * from './random-background';
