+++ src/utils/utils.test.ts
@@ -1,13 +1,12 @@
-import { DIRECTION } from '../const/direction-button';
+import { DIRECTION, ITEM_ON_PAGE } from '../const';
 import { describe } from 'vitest';
 import { getCurrentPage } from './get-current-page';
-import { defaultValueApi, itemOnPage } from './api-default-value';
 import { agreeValidate, dateValidate, fileValidate } from './validate';
 
 describe('Test utils', () => {
   const page = 1;
   const array4 = ['t', 'e', 's', 't'];
-  const array15 = [...Array(itemOnPage).keys()];
+  const array15 = [...Array(ITEM_ON_PAGE).keys()];
   const array32 = [...Array(33).keys()];
 
   const yesterday = String(new Date().setDate(new Date().getDate() - 1));
@@ -16,15 +15,6 @@ describe('Test utils', () => {
     expect(getCurrentPage(DIRECTION.NEXT, page, array4 as string[])).toBe(undefined);
     expect(getCurrentPage(DIRECTION.NEXT, page, array15 as number[])).toBe(2);
     expect(getCurrentPage(DIRECTION.PREV, page + 1, array32 as number[])).toBe(1);
-  });
-
-  it('check default values api', async () => {
-    expect(defaultValueApi(page)).toBe(
-      'https://shikimori.one/api/animes?limit=16&page=1&censored=true&score=8&order=popularity'
-    );
-    expect(defaultValueApi(page, 'test')).toBe(
-      'https://shikimori.one/api/animes?limit=16&page=1&censored=true&order=popularity&search=test'
-    );
   });
 
   it('check date validate', async () => {
